refactor: moving validation to set form component
