+++ front/src/components/sets/EditSetDialog.tsx
@@ -6,7 +6,6 @@ import { useEffect, useState } from "react";
 import useSession from "../../hooks/useSession";
 import useSetService from "../../api/set.service";
 import SetForm from "./SetForm";
-import { validateReps, validateWeight } from "../../utils/setValidators";
 import { toast } from 'react-toastify';
 import useSessionService from '../../api/session.service';
 import { Session } from '../../types/session.types';
@@ -32,7 +31,7 @@ const EditSetDialog = ({
     weight: ""
   })
 
-  const [formValid, setFormValid] = useState(false);
+  // const [formValid, setFormValid] = useState(false);
   const [submitting, setSubmitting] = useState(false);
 
   const { curSession, setCurSession } = useSession();
@@ -87,34 +86,6 @@ const EditSetDialog = ({
     
   }
 
-  const handleChange = (field: string, value: string) => {
-    setFormData({ ...formData, [field]: value });
-  }
-
-  useEffect(() => {
-    // early return if reps is null or undefined.
-    if (!formData.reps) return;
-    const errorMessage = validateReps(formData.reps);
-    setFormErrors(prev => ({ ...prev, reps: errorMessage }));
-  }, [formData.reps]);
-
-  useEffect(() => {
-    // early return if weight is null
-    if (!formData.weight) return;
-    const errorMessage = validateWeight(formData.weight);
-    setFormErrors(prev => ({ ...prev, weight: errorMessage }));
-  }, [formData.weight]);
-
-  useEffect(() => {
-    if (!formData.reps || !formData.weight) {
-      setFormValid(false);
-      return;
-    } 
-    const valid = Object.values(formErrors).every((error) => error === "");
-    setFormValid(valid);
-  }, [formErrors, formData]);
-  
-
   return (
     <Dialog
       open={open}
@@ -124,9 +95,9 @@ const EditSetDialog = ({
       <DialogContent>
         <SetForm 
           formData={formData}
+          setFormData={setFormData}
           formErrors={formErrors}
-          formValid={formValid}
-          handleChange={handleChange}
+          setFormErrors={setFormErrors}
           handleSubmit={handleSubmit}
           onClose={onClose}
           submitting={submitting}
