+++ front/src/components/sets/SetForm.tsx
@@ -2,6 +2,8 @@ import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
 import InputAdornment from '@mui/material/InputAdornment';
 import TextField from '@mui/material/TextField';
+import { Dispatch, SetStateAction, useEffect, useState } from 'react';
+import { validateReps, validateWeight } from '../../utils/setValidators';
 
 export type SetFormFields = {
   reps: string,
@@ -10,23 +12,52 @@ export type SetFormFields = {
 
 interface SetFormProps {
   formData: SetFormFields,
-  formErrors: SetFormFields
-  handleChange: (field: string, value: string) => void,
+  setFormData: Dispatch<SetStateAction<SetFormFields>>,
+  formErrors: SetFormFields,
+  setFormErrors: Dispatch<SetStateAction<SetFormFields>>,
   handleSubmit: (e: React.SyntheticEvent) => void,
-  formValid: boolean,
   submitting: boolean,
   onClose: () => void,
 }
 
 const SetForm = ({
   formData,
+  setFormData,
   formErrors,
-  handleChange,
+  setFormErrors,
   handleSubmit,
-  formValid,
   submitting,
   onClose
 }: SetFormProps) => {
+  const [formValid, setFormValid] = useState(false);
+
+  const handleChange = (field: string, value: string) => {
+    setFormData({ ...formData, [field]: value });
+  }
+
+  useEffect(() => {
+    // early return if reps is null or undefined.
+    if (!formData.reps) return;
+    const errorMessage = validateReps(formData.reps);
+    setFormErrors(prev => ({ ...prev, reps: errorMessage }));
+  }, [formData.reps, setFormErrors]);
+
+  useEffect(() => {
+    // early return if weight is null
+    if (!formData.weight) return;
+    const errorMessage = validateWeight(formData.weight);
+    setFormErrors(prev => ({ ...prev, weight: errorMessage }));
+  }, [formData.weight, setFormErrors]);
+
+  useEffect(() => {
+    if (!formData.reps || !formData.weight) {
+      setFormValid(false);
+      return;
+    } 
+    const valid = Object.values(formErrors).every((error) => error === "");
+    setFormValid(valid);
+  }, [formErrors, formData]);
+
   return (
     <Box
       component="form"
